Storage-Vitals Data from Max30105
In this code, the vital signal data (heart rate and blood oxygen) are written to an SD card using the SD.h library. The writeDataToFile() function is used to write the data to a file named vital_sign_data.txt. The function opens the file in "append" mode and writes the data to it. If the file is not opened successfully, an error message is printed to the Serial Monitor.