share data from the wearable watch- BT and WiFi
In this code, the Bluetooth communication is established using the SoftwareSerial library, and the vital signal and battery data are sent to the mobile app via Bluetooth using the btSerial.print() and btSerial.println() functions.

The WiFi communication is established using the WiFi library, and the vital signal and battery data are sent to a server using HTTP POST requests. The client.connect() function is used to establish a connection to the server, and the data is sent using the client.println() function.