wearable watch Data to Google Firebase
Firebase communication is established using the FirebaseESP32 library. The Firebase.begin() function is used to initialize the Firebase library, and the Firebase.setString() function is used to send the data to Firebase. The path parameter specifies the location where the data should be stored in Firebase, and the data parameter is a JSON string that contains the vital signal and battery data.